requirements.txt
streamlit
diffusers
transformers
torch
pillow
accelerate
pyngrok
